Refactor: 상품 도메인 v2 기준으로 리팩토링
